Update endpoint response

+++ api/src/main/java/com/mapofzones/calculations/repository/entities/IbcTransferChart.java
@@ -6,7 +6,6 @@ import lombok.Setter;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.math.BigInteger;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -25,7 +24,7 @@ public class IbcTransferChart {
 
         private String zone;
         private BigInteger totalIbcTransfersCount;
-        private Integer totalPending;
+        private BigInteger totalPending;
         private List<ChartItem> chart;
 
         @Getter
